requirements.txt
qdrant-client==1.7.3
sentence-transformers==2.3.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0

# LLM providers (optional - install what you need)
# openai>=1.12.0        # For OpenAI GPT models
# anthropic>=0.18.0     # For Anthropic Claude
# ollama>=0.1.0         # For local Ollama models
